Fixed generating a single `<name>.proto` for a Protobuf package that is composed of many `*.proto` files
In your project, we have multiple source files that all map to the same Protobuf package. For example:

- `proto/sf/substreams/v1`
  - `modules.proto`
  - `clock.proto`
  - `substreams.proto`

When using `proto-gen-prost` with:

```
protoc -I./proto --prost_out=src/pb sf/substreams/v1/substreams.proto
```

The plugin was actually exiting with a 0 exit code but no files was generated. The reason this is happening because `entry.with_output_filename(filename)` call is made only when creating the initial entry for a package, which fails if the very first file that is creating the entry (https://github.com/neoeinstein/protoc-gen-prost/blob/fe8e21a9d319c305cda0cfddd146ccddc73d36dd/protoc-gen-prost/src/lib.rs#L82-L88) is not the actual files passed to `protoc` binary.

To fix this, the condition that sets the actual `output_filename` is now check on every files that is part of a module which fixes the problem when the file to be generated is not the first file in the module's source proto files.