feat(prost-crate): add `package_separator` option
This option allows to customize the replacement character for `.` in
feature names. Those are made using proto package names, but those
contains `.`, which is not a valid character for a feature flag on
crates.io.

Currently, the dot is replaced by an `_` character, which can lead to
conflicts with similarily named packages (such as `foo.bar` and `foo_bar`).
A solid choice to avoid both issues is to use `-`, which is allowed by
crates.io but illegal in a proto package name.

With this option, an user can choose:
- to keep the current (default) behavior or `.` -> `_`
- to replace with a dash for both conflict-free and crates.io
  compatibility
- to replace with a dot (i.e. no-op) for better developer ergonomics if
  publishing to crates.io is not required