Update documentation for prost remote plugins
These plugins are now maintained and automatically updated by the buf
team at https://github.com/bufbuild/plugins. Update the documentation to
point to the new plugin locations and use the new syntax to invoke
remote plugins, as the previous plugins alpha is deprecated and going
away on July 10th.